aggiornamento docker-compose.yml

+++ hdfs/config/bootstrap.sh
@@ -3,22 +3,26 @@
 
 service ssh start
 
-#if [[ "${HDFS_MODE}" == "master" ]]; then
-#	echo "Start nodo master";
-#	hdfs namenode -format; # delete all content of hdfs
-#	$HADOOP_HDFS_HOME/sbin/start-dfs.sh;
-#	hdfs dfs -mkdir input
-#	hdfs dfs -mkdir output
-#	hdfs dfs -cp ../../data input/
-#	hdfs dfs -chown spark:spark /output
-#	echo "Nodo master configurato"
-#fi
+if [[ -n "${HDFS_WORKERS}" ]]; then
+	IFS=',' read -ra WORKERS <<< "${HDFS_WORKERS}"
+	for worker in "${WORKERS[@]}"; do
+		echo $worker >> $HADOOP_HOME/etc/hadoop/workers;
+	done
+fi
 
-hdfs namenode -format; # delete all content of hdfs
-$HADOOP_HDFS_HOME/sbin/start-dfs.sh;
-hdfs dfs -mkdir input
-hdfs dfs -mkdir output
-hdfs dfs -chown spark:spark /output
-hdfs dfs -chown nifi:nifi /input
+if [[ "${WHOAMI}" == "master" ]]; then
+	echo "Start nodo master";
+	hdfs namenode -format; # delete all content of hdfs
+	$HADOOP_HDFS_HOME/sbin/start-dfs.sh;
+	hdfs dfs -mkdir input
+	hdfs dfs -mkdir output
+	hdfs dfs -chown spark:spark /output
+	hdfs dfs -chown nifi:nifi /input
+	echo "Nodo master configurato"
+fi
 
-/bin/bash
+
+while true; do sleep 1000; done
+
+#Wait
+wait $!;
